Touchup hack for page break skip spacing
See viachristus/mujdeyi_kurtarmak#84

This is a touchup for the hack around upstream SILE bug 262.

https://github.com/simoncozens/sile/issues/262

The original hack output a space to force the typesetter to calculate a
position, but changes to the typesetter mean a space can now be
discarded. Using a kern object instead forces it to not be discarded and
does what the hack was originally supposed to do.

Eventually this will come out when the rest of the hack does.