Switch to absolute units as implemented upstream
I original wrote the absolute percentage system for this library, then
tweaked it and submitted it upstream. As it has now been included in the
current builds of SILE, this strips the local version out and uses the
units as they were implemented upstream.

See https://github.com/simoncozens/sile/pull/304