Add hyphenation exception for ‘doğru’
Fixes viachristus/mujdeyi_kurtarmak#93

Normally this wouldn't be hyphenated, but the combination of being
longer (not 3 letters) and not being hyphenatable is actually pushing
things out of bounds in a perfect storm of bad metrics.

I figure it's better to do bogus hyphenate even on the soft consonant
than stick out of a line.