Fudge page size so font sizes translate to screen
See viachristus/mk_kitapciklari#34

This is an arbitrary page size 33% bigger than a 5" screen. The idea is
that smart phone fonts need to be smaller than the ones we print in
books to achieve a reasonable number of characters per line (this nets
us ~45). Rather than trying to change font sizes in all the book styles,
we're just scaling the overall output size so the existing font sizes
play out alright.