Clear at least 2 lines at start of quotes
Fixes viachristus/mk_kitapciklari#35

If we don't clear and backtrack a large enough block when starting
quotations, we can end up clearing the space, calculating margins to
match the location, then end up getting pushed to the next page by the
page builder anyway.

This hack is as dirty as a coal mine, but clearing at least a couple
lines worth of blank space before we start quotes pretty much grantees
that at least one line of the quote will start on the page ... or that
none of it will. Breaking inside is working as long as the typesetter
puts something on the previous page, or running our metric calculations
on the next page works, just not the some-of-both option.