Change aspect ratio of app PDFs
Fixes viachristus/mk_kitapciklari#44

The aspect I had was better for HDPI phones with full screen PDF
readers, but this will actually be better for people that don't get rid
of toolbars and such. Not having another aspect ratio for the cover is
a win that's worth less than perfect full srceen reading. People that
care about that are going to be using ePub anyway!