Fix frames layout to 'undo' masters layouts
Even though _this_ layout isn't a dual sided masters, this is the
exception and our class has already loaded that stuff up. Unless we port
the class to not be based on book, we need to undo the masters layout
stuff rather that just setting up the single page template.