Set a default verse referencing scheme
See viachristus/ho_yayinlari#4, viachristus/dg_kitapciklar#19,
    viachristus/aykiri_ates#2, viachristus/mk_kitapciklari#27,
    viachristus/teolojiye_giris#3, etc.

This setting is what will determine what is _normative_ for a verse
reference in a given project or book. This setting should be propagated
down unless over-ridden, then the final value used for reformatting the
content based on the BCV parser.

Options might be:

- full
- abrv
- osis

But we'll see how it actually works out.