Undo bits of 89279e7 that created a scoping race condition
As convoluted as it seems, we need to have SILE import these into
a different scope than they end up in just loading as part of the
pre-amble. The cleanup in 89279e7 was well intentioned but broke
everything. This still cleans things up a little bit but reverts back to
the old two step way of running includes.