Handle PUBDIR dependency more carefully in DRAFT mode
The $(PUBDIR) really needs to exist to run a normal build, so check it
as a dependency of all the targets that use it. However since it can be
run in DRAFT mode without this, implement the dependency check with
logic to make sure we aren't in DRAFT mode.

We don't want to run a full init just to get $(PUBDIR) in case the user
is hacking on CaSILE or their own toolkit, so separate that out as its
own target.