Divorce geometry calcs from generating covers
Having to build the cover layouts first before having complete geometry
calculations was a pain in the behind to keep track of and parse. The
prerequisites were just too complicated, so instead build a dummy file
that has very few prerequisites itself and so can be done early in the
process and even if not strictly needed without a performance penalty.

This should also keep the geometry from rebuilding too often and forcing
covers to regenerate.