chore: Drop default +raw_sile extension for now
This will likely be added back later, but at the moment the PR I've
submitted upstream and the Docker build we're doing don't include this
feature, which is part of the SILE Reader rather than the Writer.

SILE code may still be included in Markdown, but it should be wrapped
and called out as what it is, it is not automatically inferred. Change:

    This markdown with \mycmd[arg=true]{raw SILE} inline.

To:

    This markdown with `\mycmd[arg=true]{raw SILE}`{=sile} inline.