feat: Swap LGPL for AGPL license
I was initially thinking of this project as a library included in other
software. That isn't exactly the way it took shape. It's now basically
an independent tool that wraps other tools, and it's even setup to run as
a remote service (ala CI runners).

For this use case I think the terms of the GNU Affero General Public
License make much more sense than the GNU Lesser General Public License.
It's not a major shift in goals, but the specifics of the terms make
more sense for the kind of app this turned out to be.

Of course all code up to v0.2.0 will continue to be available under the
terms of the LGPL, but all future code and releases will be under the
AGPL.