refactor(rules)!: Deprecate PUBDIR mechanism
This entire mechanism is fairly unorthodox and was specifically designed
for my own personal workflows. It is particularly troublesome because it
was designed to be used using hard links, something not all filesystems
even support.

This will be replaced with something akin to autoconf's dist files list
so that final artifacts can be deployed without getting too many
intermediates mixed up in the process. This commit doesn't get a new
system working, it only rips out the old one and leaves some
placeholders.