refactor(docker): Roll our own push to GH package registry
The official GH Action for this gets worse and worse with every version,
and the old versions aren't properly supported any more.

cf. https://github.com/docker/build-push-action/issues/208

This isn't actually that hard, and with a make rule for it we have the
option to build and push locally.