fix(rules): Remove troublesome force from draft mode
The only thing this does is paper over some other bug. If our dependency
tree was correct and complete it wouldn't matter whether we were in
draft mode or not to know what to rebuild. I suspect this was in here to
make it easier to hack on CaSILE (or SILE) itself without needing all
those things in the dependency tree. However it **breaks** our make
wrapper script because it causes unknown targets to not throw errors.