refactor(docker): Stop using tagged base images
The idea is great, but because of the way Arch's package repositories
work we can't actually use these tagged images to re-create old versions
anyway. Given the build process is going to end up with the latest
packages anyway there is no reason I can think of to not just start with
the freshest image possible.