refactor(class)!: Replace open-page with open-spread
The open-page function had all sorts of edge case issues (both similar
to SILE's issues and unique to itself). This uses a completely different
mechanism with pagend hooks to track progress.

BREAKING CHANGE: The open-page function has been removed, but
open-spread is not a drop in replacement for all previous usage because
it will force an even page opening if odd is false.