chore(renderings): Default POV single execution to off
If 2 jobs kill machine, 1 probably will too. It is mostly texture memory
that is an OOM issue. On the other hand with the bulk of the rendering
pass being single threaded anyway we're loosing a lot of job run
performance on modern hardware.

Leave the option to enable for some runners, but disable by default.