feat(covers): Reduce default resource resolution to 600dpi
It's totally okay to generate stuff at 1200dpi for final press
resources, but for everyday use during book development this is a huge
memory hog and makes CI jobs harder to manage than they should be.
I think it is going to be easier to boost the value for special high
quality builds than it is to constantly deduce the default in most CI
jobs and terminals.