refactor(cli)!: Rename 'script' subcommand to 'run'
BREAKING CHANGE: The fact that subcommands are not all verbs has been
bothering me for a while. Several times I've sat down recently and had
to look up what the subcommand was. This is more along the lines of what
I seem to expect every time I'm away from it for a while. This choice
makes all the subcommands into verbs. Alternatives like 'execute' and
'do' were considered, this is just seemingly the most obvious
alternative.