refactor(rules)!: Drop BUILDDIR as a make variable and make it a config option
BREAKING CHANGE: The $(BUILDDIR) vaiable used in many make targets can
no longer by set by make rules, it must be set via the builddir option
via YAML configs or ENV vars. This makes it easier to use in scripts
that may not run as a child process of make.