fix(rules): Fix project-local font directory handling
We've been stripping the last segment for no reason (as if the value was
a file and we needed the directory part) but the value is supposed to be
a directory. This works in many cases because of the way fontconfig
traverses directories, but creates priority issues when system paths for
the same fonts have less recursion.