feat(deps): Make @Omikhleia's ptable package available by default
It seems like I use this on most of my projects these days. Now that
silex (which it depends on, so becomes a transitive dependency here too)
doesn't hijack SILE's namespace by default, there isn't any reason not
to provide this out of the box. So instead of having them have to
install it in their own Docker images or Lua Rockspec, lets just bundle
it by default.